Create Sudoku Solver
Description
Sudoku is a logical based, combinatorial number-placement puzzle. The objective is to fill a 9×9 grid with digits so that each column, each row, and each of the nine 3×3 subgrids that compose the grid contain all of the digits from 1 to 9.

For example:

Image

The puzzle is represented using a 9 by 9 structure array (struct element grid[9][9]), and each element's num is an integer between 0~9 (0 means there is no value yet, corresponding to the white area in the figure, and can_modify is set to true). The elements that you should not change are the gray shaded area in the figure and they belong to this problem (can_modify is set to false).

Implement the solve() function that can automatically solve a given sudoku puzzle (you should directly modify grid[9][9] and fill in the correct answer).

Note: There is exactly one solution for each testcase.

數獨是一個邏輯和數字組合的遊戲。其目標是將數字填入 9x9 的格子內，並使其每一行，列，和 3x3 的格子皆由 1～9 的數字所組成。

例如：如上圖

本次題目中，數獨由一個 9x9 的 struct array 表示 (struct element grid[9][9])，而每個 element 的 num 為介於 0~9 的 integer (0 表示還沒有填入任何數字，且其 can_modify 設為 true)。不應更改圖中灰色區域的格子，這些格子為題目 (can_modify 設為 false)。

實作 solve() 函數來自動解出給定的數獨 (直接修改 grid[9][9] 即可)。

註：每個測資都會是唯一解。

Input
The input is a sudoku puzzle. The elements, which you should fill in, are represented by '.'.

Output
The output is the solved sudoku puzzle.

Limits
Your program needs to finish task in 3 seconds.

Your program can only use memory less than 20000 KB.