Create Throwing Cards Away
Description
Given an ordered deck of n cards numbered 1 to n with card 1 at the top and card n at the bottom, throw away the top card and move the next card that is on the top of the deck to the bottom of the deck. Repeat this process m times and find the sequence of discarded cards.

給一副排組，由上而下的編號為 1~n，共會進行 m 次抽取，每次抽走最上層的牌並輸出此牌的編號，再將最上層的牌放到牌組的最下方。

Input
Two integers n & m.

Output
The sequence of the discarded cards. Every number is followed by one space.

Limits
Your program needs to finish task in 1 second.

Your program can only use memory less than 65535 KB.